Can delete transactions

+++ src/routes/api/transaction/+server.ts
@@ -7,39 +7,39 @@ export const GET: RequestHandler = async function ({ platform }) {
 }
 
 export const POST: RequestHandler = async function ({ request, platform }) {
-  const { employee_id, dollars, cents, tyear, tmonth, tday, thour, tminute } = await request.json();
+  const { op } = await request.json();
 
-  try {
-    await platform?.env.DB.prepare('INSERT INTO transactions (id, employee_id, dollars, cents, tyear, tmonth, tday, thour, tminute) VALUES (NULL, ?, ?, ?, ?, ?, ?, ?, ?)')
-                            .bind(employee_id, dollars, cents, tyear, tmonth, tday, thour, tminute)
-                            .run();
-    
+  if(op.action == "insert") {
+    try {
+      await platform?.env.DB.prepare('INSERT INTO transactions (id, employee_id, dollars, cents, tyear, tmonth, tday, thour, tminute) VALUES (NULL, ?, ?, ?, ?, ?, ?, ?, ?)')
+                              .bind(op.data.employee_id, op.data.dollars, op.data.cents, op.data.tyear, op.data.tmonth, op.data.tday, op.data.thour, op.data.tminute)
+                              .run();
+      
       return json({
         success: true
       });
 
-  } catch (err) {
-    return new Response(err, {
-      status: 500
-    });
+    } catch (err) {
+      return new Response(err, {
+        status: 500
+      });
+    }
+  }
+
+  if(op.action == "delete") {
+    try {
+      await platform?.env.DB.prepare('DELETE FROM transactions WHERE id = ?')
+                              .bind(op.data.transaction_id)
+                              .run();
+  
+      return json({
+        success: true
+      });
+  
+    } catch (err) {
+      return new Response(err, {
+        status: 500
+      });
+    }
   }
 };
-
-export const DELETE: RequestHandler = async function ({ request, platform }) {
-  const { transaction_id } = await request.json();
-
-  try {
-    await platform?.env.DB.prepare('DELETE FROM transactions WHERE id = ?')
-                            .bind(transaction_id)
-                            .run();
-
-      return json({
-        success: true
-      });
-
-  } catch (err) {
-    return new Response(err, {
-      status: 500
-    });
-  }
-}
